requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
bcrypt==4.3.0
    # via -r requirements.in
build==1.3.0
    # via pip-tools
cffi==1.17.1
    # via cryptography
click==8.2.1
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
cryptography==45.0.6
    # via python-jose
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.116.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
packaging==25.0
    # via build
passlib==1.7.4
    # via -r requirements.in
pip-tools==7.5.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.35.0
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
